+++ src/04-sentence/links.js
@@ -1,4 +1,5 @@
 // const helpers = require('../lib/helpers');
+const parse_interwiki = require('./interwiki');
 const ignore_links = /^:?(category|catégorie|Kategorie|Categoría|Categoria|Categorie|Kategoria|تصنيف|image|file|image|fichier|datei|media|special|wp|wikipedia|help|user|mediawiki|portal|talk|template|book|draft|module|topic|wiktionary|wikisource):/i;
 const external_link = /\[(https?|news|ftp|mailto|gopher|irc)(:\/\/[^\]\| ]{4,1500})([\| ].*?)?\]/g;
 const link_reg = /\[\[(.{0,120}?)\]\]([a-z']+)?(\w{0,10})/gi; //allow dangling suffixes - "[[flanders]]'s"
@@ -15,6 +16,8 @@ const external_links = function(links, str) {
   });
   return links;
 };
+
+
 
 const internal_links = function(links, str) {
   //regular links
@@ -48,9 +51,9 @@ const internal_links = function(links, str) {
       obj.anchor = b;
       return '';
     });
-    // let anchor = obj.page.match(/#[^ ]{1,100}/);
-    // if(anchor!==null){
-    // }
+    //grab any fr:Paris parts
+    obj = parse_interwiki(obj);
+
     if (txt !== null && txt !== obj.page) {
       obj.text = txt;
     }